Harden portal routes with customer actor helpers
